backend/requirements.txt.txt
flask
pycryptodome
google-auth
google-auth-oauthlib
google-auth-httplib2
googleapiclient
